+++ app/src/main/java/com/example/yourgoldenhour/navigation/Screen.kt
@@ -7,10 +7,10 @@ sealed class Screen(val route: String) {
     object AddPhotoScreen : Screen("camera")
 
     object PhotoScreen : Screen("photo_screen/{photoId}"){
-        fun createRoute(photoId: Int): String = "photo_screen/$photoId"
+        fun createRoute(photoId: Long): String = "photo_screen/$photoId"
     }
 
     object EditScreen : Screen("edit_screen/{photoId}"){
-        fun createRoute(photoId: Int): String = "edit_screen/$photoId"
+        fun createRoute(photoId: Long): String = "edit_screen/$photoId"
     }
 }
